requirements.txt
Django>=4.0,<5.0
Pillow>=9.0.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
mysqlclient>=2.1
black>=25.0.0
flake8>=7.0.0
